// src/generated/api/models/ResponseDtoProfileImageResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProfileImageResponse } from './ProfileImageResponse';
import {
    ProfileImageResponseFromJSON,
    ProfileImageResponseFromJSONTyped,
    ProfileImageResponseToJSON,
    ProfileImageResponseToJSONTyped,
} from './ProfileImageResponse';

/**
 * 
 * @export
 * @interface ResponseDtoProfileImageResponse
 */
export interface ResponseDtoProfileImageResponse {
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoProfileImageResponse
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponseDtoProfileImageResponse
     */
    message?: string;
    /**
     * 
     * @type {ProfileImageResponse}
     * @memberof ResponseDtoProfileImageResponse
     */
    data?: ProfileImageResponse;
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoProfileImageResponse
     */
    timestamp?: number;
}

/**
 * Check if a given object implements the ResponseDtoProfileImageResponse interface.
 */
export function instanceOfResponseDtoProfileImageResponse(value: object): value is ResponseDtoProfileImageResponse {
    return true;
}

export function ResponseDtoProfileImageResponseFromJSON(json: any): ResponseDtoProfileImageResponse {
    return ResponseDtoProfileImageResponseFromJSONTyped(json, false);
}

export function ResponseDtoProfileImageResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseDtoProfileImageResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'message': json['message'] == null ? undefined : json['message'],
        'data': json['data'] == null ? undefined : ProfileImageResponseFromJSON(json['data']),
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function ResponseDtoProfileImageResponseToJSON(json: any): ResponseDtoProfileImageResponse {
    return ResponseDtoProfileImageResponseToJSONTyped(json, false);
}

export function ResponseDtoProfileImageResponseToJSONTyped(value?: ResponseDtoProfileImageResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'message': value['message'],
        'data': ProfileImageResponseToJSON(value['data']),
        'timestamp': value['timestamp'],
    };
}